K-Means/requirements.txt
matplotlib 
numpy 
scikit-datasets 
Pillow